created new ticket booking window and migrated dashboard booking to new window.
